requirements.txt
numpy~=1.23.4
Pillow~=9.2.0
requests~=2.28.1
